lcd comm

+++ bbridge/rtkPosition.h
@@ -6,6 +6,6 @@
 int init_rtklib_client(void);
 int dismiss_rtklib_client(void);
 void * read_frm_RTKlib(void);
-float get_latitude(void);
-float get_longitude(void);
-float get_height(void);
+long long get_latitude(void);
+long long get_longitude(void);
+long long get_height(void);
